remove db_creds
