+++ .gitignore
@@ -1,5 +1,4 @@
 out
 node_modules
 .vscode-test/
-db_creds.json
 *.vsix
