--- db_creds.json
@@ -1,16 +1,0 @@
-{
-    "current_db": "default_db",
-    "dbs": {
-        "default_db": {
-            "type": "athena"
-        },
-        "personal_db": {
-            "type": "postgres",
-            "host": "",
-            "port": "",
-            "database": "",
-            "user": "",
-            "password": ""
-        }
-    }
-}
