+++ src/utils.ts
@@ -1,7 +1,3 @@
-import fs = require('fs');
-import path = require('path');
-
-
 export function wait(ms: number){
     // Synchronous wait for a number of milliseconds.
 	var start = new Date().getTime();
@@ -28,7 +24,7 @@ export function parse_queries(query: string) {
 	
 	var queries = query.split(';');
 	var parsed_queries = [];
-	const query_start_regex = /[Ss][Ee][Ll][Ee][Cc][Tt]/;
+	const query_start_regex = new RegExp("([Ww][Ii][Tt][Hh][^;]+)|([De][Ee][Ll][Ee][Tt][Ee][^;]+)|([Cc][Oo][Pp][Yy][^;]+)|([Ii][Nn][Ss][Ee][Rr][Tt][^;]+)|([Uu][Pp][Dd][Aa][Tt][Ee][^;]+)|([gG][rR][aA][nN][tT][^;]+)|([aA][lL][tT][eE][rR][^;]+)|([dD][rR][oO][pP][^;]+)|([cC][rR][eE][aA][tT][eE][^;]+)|([sS][eE][lL][eE][cC][tT][^;]+)");
 
 	for (var i = 0; i < queries.length; i++) {
 
@@ -43,16 +39,11 @@ export function parse_queries(query: string) {
 	return parsed_queries
 }
 
-export function getDbJson() {
-	const filePath = path.join(__dirname, '../db_creds.json')
-	const dbJson = JSON.parse(fs.readFileSync(filePath, 'utf8'));
+export function getNthColumn(table: any, n: number) {
+	var col = [];
+	for (var i = 0; i < table.length; i++) {
+		col.push(table[i][n])
+	}
 
-	return dbJson
-}
-
-export function getCurrentDbType() {
-	const dbJson = getDbJson()
-	const currentDb = dbJson['current_db']
-
-	return dbJson['dbs'][currentDb]['type']
+	return col
 }
